rename LoginFlow to LoginModal
